# src/utils/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API para gestão de fazendas",
    "description": "Essa API tem como objetivo a gestão de fazendas",
    "termsOfService": "http://localhost:3001/terms",
    "contact": {
      "email": "rafaelrabelodev@gmail.com"
    },
    "version": "1.0.0",
    "x-raw-content": {
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/",
      "description": "API para teste"
    }
  ],
  "tags": [
    {
      "name": "Users",
      "description": "Operações relacionadas a usuários"
    },
    {
      "name": "Accounts",
      "description": "Operações relacionadas a contas"
    },
    {
      "name": "Authentication",
      "description": "Operações relacionadas a autenticação"
    }
  ],
  "paths": {
    "/auth/signin": {
      "post": {
        "summary": "Autenticação de Usuário",
        "description": "Realiza a autenticação de um usuário.",
        "requestBody":{
          "description": "Credenciais de autenticação",
          "required": true,
          "content": {
            "application/json":{
              "example": {
                "email": "exemplo@email.com",
                "password": "senha123"
              }
            }
          } 
        },
        "responses": {
          "200": {
            "description": "Autenticação bem-sucedida. Retorna um token de acesso."
          },
          "401": {
            "description": "Falha na autenticação. Usuário não encontrado ou senha incorreta."
          }
        },
        "tags":[
          "Authentication"
        ]
      }
    },
    "/users": {
      "get": {
        "summary": "Lista de Usuários",
        "description": "Obtém a lista de todos os usuários.",
        "responses": {
          "200": {
            "description": "Lista de usuários obtida com sucesso."
          }
        },
        "tags": [
          "Users"
        ]
      },
      "post": {
        "summary": "Criar Usuário",
        "description": "Cria um novo usuário.",
        "requestBody": {
          "description": "Dados do usuário a ser criado",
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "name": "Exemplo de Nome",
                "email": "exemplo@email.com",
                "password": "senha123"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso."
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/accounts": {
      "get": {
        "summary": "Lista de Contas",
        "description": "Obtém a lista de todas as contas.",
        "responses": {
          "200": {
            "description": "Lista de contas obtida com sucesso."
          }
        },
        "tags": [
          "Accounts"
        ]
      },
      "post": {
        "summary": "Criar Conta",
        "description": "Cria uma nova conta.",
        "requestBody": {
          "description": "Dados da conta a ser criada",
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "user_id": 1,
                "name": "Exemplo de Conta"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Conta criada com sucesso."
          }
        },
        "tags": [
          "Accounts"
        ]
      }
    },
    "/accounts/{id}": {
      "get": {
        "summary": "Buscar Conta por ID",
        "description": "Obtém uma conta pelo seu ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID da conta a ser buscada."
          }
        ],
        "responses": {
          "200": {
            "description": "Conta obtida com sucesso."
          }
        },
        "tags": [
          "Accounts"
        ]
      },
      "put": {
        "summary": "Atualizar Conta",
        "description": "Atualizar uma conta.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID da conta a ser atualizada."
          }
        ],
        "requestBody": {
          "description": "Dados da conta a ser atualizada",
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "user_id": 1,
                "name": "Conta Atualizada"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Conta atualizada com sucesso."
          }
        },
        "tags": [
          "Accounts"
        ]
      },
      "delete": {
        "summary": "Deletar Conta",
        "description": "Deleta uma conta pelo seu ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID da conta a ser atualizada."
          }
        ],
        "responses": {
          "200": {
            "description": "Conta deletada com sucesso."
          }
        },
        "tags": ["Accounts"]
      }
    }
  }
}